% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paralleltest.R
\name{parallel_test}
\alias{parallel_test}
\title{Title parallel test}
\usage{
parallel_test(data, bd, br, j, snc)
}
\arguments{
\item{data}{y: Response value;
x2: Dilution Dose;
x1: the group number - 0 for the standard group, 1,2,3... for the test group}

\item{bd}{the base number of logarithmic transformation of dilution dose}

\item{j}{the number of test group}

\item{snc}{Standard Nominal Concentration}

\item{bp}{the base number of logarithelphmic transformation of relative potency}
}
\value{
a list
}
\description{
Title parallel test
}
\examples{
result <- parallel_test(data=mydata, bd=2, br=2, j=1, snc=48000000)

}
